Adding batch inference.

+++ src/chatterbox/models/t3/modules/cond_enc.py
@@ -67,7 +67,7 @@ class T3CondEnc(nn.Module):
             "no embeddings for cond_prompt_speech_tokens"
 
         # Speaker embedding projection
-        cond_spkr = self.spkr_enc(cond.speaker_emb.view(-1, self.hp.speaker_embed_size))[:, None]  # (B, 1, dim)
+        cond_spkr = self.spkr_enc(cond.speaker_emb.view(-1, self.hp.speaker_embed_size).to(self.spkr_enc.weight.dtype))[:, None]  # (B, 1, dim)
         empty = torch.zeros_like(cond_spkr[:, :0])  # (B, 0, dim)
 
         # TODO CLAP
@@ -85,7 +85,7 @@ class T3CondEnc(nn.Module):
         cond_emotion_adv = empty  # (B, 0, dim)
         if self.hp.emotion_adv:
             assert cond.emotion_adv is not None
-            cond_emotion_adv = self.emotion_adv_fc(cond.emotion_adv.view(-1, 1, 1))
+            cond_emotion_adv = self.emotion_adv_fc(cond.emotion_adv.view(-1, 1, 1).to(self.emotion_adv_fc.weight.dtype))
 
         # Concat and return
         cond_embeds = torch.cat((
